user-data: allow missing user data.
When I deduplicated this code I mistakenly used parse_headers() in
the same way twice. This was incorrect.

We definitely need to error out if a 404 or something else happens
for the SSH keys, since then we've lost entirely.

But a missing `user-data` is mostly harmless and optional, so, in
case a 404 happens for the second `GET` request, we shouldn't error
out, because this makes the instance unusable. Instead, we can
carry on as normal.

+++ src/ucd-data-fetch.c
@@ -337,13 +337,15 @@ int main(int argc, char *argv[]) {
 
 	/* parse/discard the header and body */
 	result = parse_headers(f, &cl);
-	if (result != 1) {
+	if (result == 0) {
+		/* error - exit */
 		fclose(f);
 		close(out);
 		FAIL("parse_headers()");
 	}
 
-	if (write_lines(out, f, cl) != 0) {
+	/* don't write part #2 if 404 or some non-error */
+	if ((result != 2) && (write_lines(out, f, cl) != 0)) {
 		close(out);
 		fclose(f);
 		unlink(outpath);
